# docs/swagger.yaml
basePath: /api/v1
definitions:
  main.AddIssueAssigneePayload:
    properties:
      user_id:
        type: string
    type: object
...
  main.CreateIssuePayload:
    properties:
      description:
        type: string
      label_ids:
        description: LabelIDs are applied in the same transaction; each must belong
          to the project.
        items:
          type: integer
        type: array
      priority:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
  main.UpdateIssuePayload:
    properties:
      cycle_id:
        description: CycleID moves the issue into a cycle, or out of one with null.
        type: integer
      description:
        type: string
      priority:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
  store.Assignee:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  store.Issue:
    properties:
      assignees:
        items:
          $ref: '#/definitions/store.Assignee'
        type: array
      author_id:
        type: string
      created_at:
        type: string
      cycle_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      labels:
        items:
          $ref: '#/definitions/store.LabelSummary'
        type: array
      priority:
        type: string
      project_id:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
...
  store.LabelSummary:
    properties:
      color:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for Trackforge application.
  title: Trackforge API
  version: "1.0"
paths:
...
  /issues/{issueID}:
    delete:
      description: Anyone in the project. Its comments and activity trail go with
        it; setting the status to cancelled instead keeps the history.
      parameters:
      - description: Issue ID
        in: path
        name: issueID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Delete an issue
      tags:
      - issues
    get:
      parameters:
      - description: Issue ID
        in: path
        name: issueID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Issue'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Get one issue
      tags:
      - issues
    patch:
      consumes:
      - application/json
      description: Fields left out of the payload are unchanged. Every change is recorded
        on the issue's activity trail.
      parameters:
      - description: Issue ID
        in: path
        name: issueID
        required: true
        type: integer
      - description: Fields to change
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateIssuePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Issue'
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "409":
          description: Conflict
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Edit an issue
      tags:
      - issues
...
  /issues/{issueID}/assignees:
    post:
      consumes:
      - application/json
      description: The user must already be a member of the project. An issue may
        have any number of assignees.
      parameters:
      - description: Issue ID
        in: path
        name: issueID
        required: true
        type: integer
      - description: User to assign
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.AddIssueAssigneePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Assignee'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "409":
          description: Conflict
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Assign a user to an issue
      tags:
      - issues
  /issues/{issueID}/assignees/{userID}:
    delete:
      parameters:
      - description: Issue ID
        in: path
        name: issueID
        required: true
        type: integer
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Unassign a user from an issue
      tags:
      - issues
...
  /projects/{id}/issues:
    get:
      description: Issues are returned newest first, each with its assignees and labels.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Issue'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: List a project's issues
      tags:
      - issues
    post:
      consumes:
      - application/json
      description: Omitting status or priority falls back to 'backlog' and 'no-priority'.
        label_ids must all belong to the project, or nothing is created (422).
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Issue to create
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateIssuePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Issue'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Create an issue on a project
      tags:
      - issues
...
securityDefinitions:
  BearerAuth:
    description: JWT from POST /auth/login. Paste it with the scheme included, e.g.
      "Bearer eyJhbGciOi..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"